# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API medico-paciente-app",
        "description": "O objetivo desta API é listar médicos e seus pacientes, divididos por cidade.",
        "contact": {
            "email": "cleison.npi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Ambiente de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/pacientes/:id_paciente": {
            "post": {
                "tags": [
                    "atualizar pacientes"
                ],
                "summary": "Cadastra um novo paciente",
                "operationId": "bcd3ffdef67dc574a3d76ce3e3257b79",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Bearer token para autenticação",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Matheus Henrique"
                                    },
                                    "celular": {
                                        "type": "string",
                                        "example": "(11) 9 8432-5789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "paciente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Matheus Henrique"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "795.429.941-60"
                                                },
                                                "celular": {
                                                    "type": "string",
                                                    "example": "(11) 9 8432-5789"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - A autenticação é obrigatória"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}